refactor: ShapeInterpolation EditSelectionSChoice logic moved to mode
